EXSWHTEC-134 - Move GLContextScopeGuard to gl_interop_common.hh
